requirements.txt
open3d>=0.17.0
numpy>=1.21.0
pyyaml>=6.0
click>=8.0.0
tqdm>=4.64.0
pydantic>=2.0.0
pathlib2>=2.3.7
laspy>=2.0.0